move RunError to the outstation session module since that's the only place it's used anymore
