+++ dnp3/src/util/task.rs
@@ -1,5 +1,4 @@
 use crate::app::Shutdown;
-use crate::link::error::LinkError;
 
 pub(crate) struct Receiver<T> {
     inner: crate::tokio::sync::mpsc::Receiver<T>,
@@ -21,23 +20,5 @@ impl<T> Receiver<T> {
 impl From<crate::tokio::sync::oneshot::error::RecvError> for Shutdown {
     fn from(_: crate::tokio::sync::oneshot::error::RecvError) -> Self {
         Shutdown
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub(crate) enum RunError {
-    Link(LinkError),
-    Shutdown,
-}
-
-impl From<Shutdown> for RunError {
-    fn from(_: Shutdown) -> Self {
-        RunError::Shutdown
-    }
-}
-
-impl From<LinkError> for RunError {
-    fn from(err: LinkError) -> Self {
-        RunError::Link(err)
     }
 }
